requirements.txt
# asgiref==3.7.2
beautifulsoup4==4.12.3
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
crispy-bootstrap4==2023.1
Django==5.0
django-chartjs==2.3.0
django-crispy-forms==2.1
django-etc==1.4.0
django-hitcount==1.3.5
django-tinymce==4.1.0
djangorestframework==3.14.0
docopt==0.6.2
holidays==0.40
idna==3.6
# lxml==5.1.0
mysqlclient==2.2.6
num2words==0.5.13
# numpy==1.26.4
# pandas==2.2.2
# Pillow==10.1.0
# psycopg2-binary==2.9.9
pypng==0.20220715.0
Pytanggalmerah==3.1.1
# python-dateutil==2.8.2
python-decouple==3.8
# python-docx==1.1.0
# pytz==2023.3.post1
qrcode==7.4.2
reportlab==4.2.2
requests==2.31.0
# six==1.16.0
soupsieve==2.6
# sqlparse==0.4.4
# typing_extensions==4.9.0
# tzdata==2023.4
urllib3==2.1.0
zcache==1.0.2
